backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project]
name = "speciesnet-studio-backend"
version = "0.1.0"
description = "Backend for SpeciesNet Studio — a self-hostable review UI for wildlife classifier predictions"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "sqlmodel>=0.0.21",
    "alembic>=1.13.0",
    "pillow>=10.0.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
speciesnet = ["speciesnet>=5.0.0"]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "types-pyyaml>=6.0.12.20260408",
]
